feat: 🎸 controlClassOnly
