+++ package.json
@@ -54,7 +54,7 @@
     "karma-jasmine-html-reporter": "^1.5.0",
     "lint-staged": "^9.2.0",
     "ng-packagr": "^15.0.1",
-    "prettier": "^1.18.2",
+    "prettier": "2.8.8",
     "standard-version": "^6.0.1",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
